Moving directories and adding Search to the BottomNavigationBar
