+++ app/src/main/java/com/example/mychef/Previews.kt
@@ -2,7 +2,7 @@ package com.example.mychef
 
 import androidx.compose.runtime.Composable
 import androidx.compose.ui.tooling.preview.Preview
-import com.example.mychef.presentation.home.HomeScreen
+import com.example.mychef.ui.home.HomeScreen
 
 @Composable
 @Preview(showBackground = true)
