+++ app/src/main/java/com/example/mychef/ui/favorites/FavoritesScreen.kt
@@ -1,4 +1,4 @@
-package com.example.mychef.presentation.cart
+package com.example.mychef.ui.favorites
 
 import androidx.compose.foundation.layout.Box
 import androidx.compose.foundation.layout.fillMaxSize
@@ -9,13 +9,13 @@ import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
 
 @Composable
-fun CartScreen() {
+fun FavoritesScreen() {
     Box(modifier = Modifier
         .fillMaxSize(),
         contentAlignment = Alignment.Center
     ){
         Text(
-            text = "Cart Screen",
+            text = "Favorites Screen",
             style = MaterialTheme.typography.headlineLarge
         )
     }
