+++ app/src/main/java/com/example/mychef/ui/search/SearchScreen.kt
@@ -1,0 +1,22 @@
+package com.example.mychef.ui.search
+
+import androidx.compose.foundation.layout.Box
+import androidx.compose.foundation.layout.fillMaxSize
+import androidx.compose.material3.MaterialTheme
+import androidx.compose.material3.Text
+import androidx.compose.runtime.Composable
+import androidx.compose.ui.Alignment
+import androidx.compose.ui.Modifier
+
+@Composable
+fun SearchScreen() {
+    Box(modifier = Modifier
+        .fillMaxSize(),
+        contentAlignment = Alignment.Center
+    ){
+        Text(
+            text = "Search Screen",
+            style = MaterialTheme.typography.headlineLarge
+        )
+    }
+}
